Add files via upload
This notebook contains a simulation of three efficient algorithms on a multiarmed bandit problem in comparison to an exploit and explore only strategy. 

The story for the simulation explores an agent that visits three restaurants that have different standard deviations and means that are unknown to the player. 
By striking a balance between exploration and exploitation the agent tries to optimize his strategy to get the highest reward. 
In the end, the amount of restaurants and the corresponding standard deviations and means are extended to increase the amount of uncertainty. 

It shown through this example, that the efficient algorithms, namely, the epsilon greedy, upper confidence bounds and the thompson sampling all perform well under uncertainty.