Add files via upload
This notebook contains the first part of our machine learning challenge when trying recreate the compass score used in american courts by building a binary classifier to predict a person's chance of recidivism. In particular this notebook contains data exploration & visulization, data cleaning, preprocessing and feature engineering for the ML models. 
The results of the project will be presented in preceding notebooks.