Add files via upload
This notebook is part two of the machine learning challenge to build a binary classifier for the Compass algorithm. It deals with linear models, namely a logistic regression and Ridge regression and a KNN classifier for prediction. The results support the findings made by ProPublica, that the Compass score is unjustified racially biased against African Americans. We argue, by reversing the challenge to predict the race out of the given features, that there exists a causal effect of race on seemingly non-racial features that are important for making a prediction. As such we strongly recommend against the use of a binary classifier in court since we can't control for the racial bias at the source of data collection.