Add files via upload
This notebook is the last part of the machine learning challenge to build a binary classifier for the Compass algorithm. It deals with Boosting as a machine learning model. The results support the findings made by ProPublica and findings of the prior notebook dealing with linear models, that the Compass score is unjustified racially biased against African Americans. We again argue, by reversing the challenge to predict the race out of the given features, that there exists a causal effect of race on seemingly non-racial features that are important for making a prediction. As such we strongly recommend against the use of a binary classifier in court since we can't control for the racial bias at the source of data collection.